+++ fastly/ngwaf/v1/virtualpatches/doc.go
@@ -1,0 +1,3 @@
+// Package virtualpatches contains API operations to get, list and
+// patch Fastly Next-Gen WAF virtual patches
+package virtualpatches