+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Breaking:
 
 ### Enhancements:
+- feat(ngwaf): add support for virtualpatches ([#680](https://github.com/fastly/go-fastly/pull/680))
 
 - feat(ngwaf/events): add support for the events portion of the ngwaf api ([#684](https://github.com/fastly/go-fastly/pull/684))
 - feat(fastly/ngwaf/v1/redactions): adds CRUD operations for the redactions portion of the ngwaf API ([#685](https://github.com/fastly/go-fastly/pull/685))